refactor: rename core initializer file
