+++ tests/engine/coreBuilder.test.ts
@@ -4,7 +4,7 @@ import { gameEngineToken } from '@core/gameEngine'
 import { turnSchedulerToken } from '@core/turnScheduler'
 import { inputMatrixBuilderToken } from '@builders/inputMatrixBuilder'
 import { engineInitializerToken } from '@core/initializers/engineInitializer'
-import { coreInitializerToken } from '@core/initializers/coreInitializers'
+import { coreInitializerToken } from '@core/initializers/coreInitializer'
 import { engineStartInitializerToken } from '@core/initializers/engineStartInitializer'
 import { providersInitializerToken } from '@core/initializers/providersInitializer'
 import { managersInitializerToken } from '@core/initializers/managersInitializer'
